utils/requirements.txt
beautifulsoup4==4.11.1
python-dotenv==1.0.1
selenium==4.23.1
webdriver_manager==3.7.0
